feat: create conditional tooltip for chips
