+++ src/core_modules/capture-core/components/ConditionalTooltipForChip/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { ConditionalTooltipForChip } from './ConditionalTooltipForChip.component';
