+++ src/core_modules/capture-core/components/TemplateSelector/TemplateSelectorChip.component.js
@@ -1,33 +1,16 @@
 // @flow
 import React, { useCallback } from 'react';
-import { Chip, Tooltip } from '@dhis2/ui';
-import { withStyles } from '@material-ui/core/styles';
+import { Chip } from '@dhis2/ui';
 import type { WorkingListTemplate } from './workingListsBase.types';
+import { ConditionalTooltipForChip } from '../ConditionalTooltipForChip';
 
 type Props = {
     template: WorkingListTemplate,
     onSelectTemplate: (template: WorkingListTemplate) => void,
-    ...CssClasses,
 };
 
-const styles = {
-    // button style reset
-    button: {
-        border: 'none',
-        backgroundColor: 'transparent',
-        borderRadius: '16px',
-        padding: 0,
-        margin: 0,
-        minWidth: 0,
-        minHeight: 0,
-        '&:hover': {
-            backgroundColor: 'transparent',
-        },
-    },
-};
-
-const TemplateSelectorChipPlain = (props: Props) => {
-    const { template, onSelectTemplate, classes } = props;
+export const TemplateSelectorChip = (props: Props) => {
+    const { template, onSelectTemplate } = props;
     const { displayName } = template;
 
     const selectTemplateHandler = useCallback(() => {
@@ -37,35 +20,23 @@ const TemplateSelectorChipPlain = (props: Props) => {
     const text = displayName.length > 30 ? `${displayName.substring(0, 27)}...` : displayName;
 
     return (
-        <Tooltip
+        <ConditionalTooltipForChip
             content={displayName}
             placement={'top'}
             openDelay={800}
+            enabled={displayName.length > 30}
+            onClick={selectTemplateHandler}
         >
-            {({ ref, onMouseOver, onMouseOut }) => (
-                <button
-                    ref={ref}
-                    onClick={selectTemplateHandler}
-                    className={classes.button}
-                    onMouseOver={onMouseOver}
-                    onMouseOut={onMouseOut}
-                    onFocus={onMouseOver}
-                    onBlur={onMouseOut}
-                >
-                    <Chip
-                        marginTop="0"
-                        marginBottom="0"
-                        marginLeft="0"
-                        marginRight="0"
-                        dataTest="workinglist-template-selector-chip"
-                        onClick={selectTemplateHandler}
-                    >
-                        {text}
-                    </Chip>
-                </button>
-            )}
-        </Tooltip>
+            <Chip
+                marginTop="0"
+                marginBottom="0"
+                marginLeft="0"
+                marginRight="0"
+                dataTest="workinglist-template-selector-chip"
+            >
+                {text}
+            </Chip>
+        </ConditionalTooltipForChip>
+
     );
 };
-
-export const TemplateSelectorChip = withStyles(styles)(TemplateSelectorChipPlain);
